+++ packages/firestore/test/unit/specs/spec_builder.ts
@@ -54,7 +54,6 @@ import {
   SpecWriteAck,
   SpecWriteFailure
 } from './spec_test_runner';
-import { PlatformSupport } from '../../../src/platform/platform';
 
 // These types are used in a protected API by SpecBuilder and need to be
 // exported.
@@ -952,21 +951,18 @@ export class SpecBuilder {
         // `query` is not added yet.
         this.activeTargets[targetId] = {
           queries: [SpecBuilder.queryToSpec(query), ...activeQueries],
-          // Convert to base64 string so it can later be parsed into ByteString.
-          resumeToken: PlatformSupport.getPlatform().btoa(resumeToken || '')
+          resumeToken: resumeToken || ''
         };
       } else {
         this.activeTargets[targetId] = {
           queries: activeQueries,
-          // Convert to base64 string so it can later be parsed into ByteString.
-          resumeToken: PlatformSupport.getPlatform().btoa(resumeToken || '')
+          resumeToken: resumeToken || ''
         };
       }
     } else {
       this.activeTargets[targetId] = {
         queries: [SpecBuilder.queryToSpec(query)],
-        // Convert to base64 string so it can later be parsed into ByteString.
-        resumeToken: PlatformSupport.getPlatform().btoa(resumeToken || '')
+        resumeToken: resumeToken || ''
       };
     }
   }