+++ packages/firestore/CHANGELOG.md
@@ -1,4 +1,8 @@
 # Unreleased
+- [fixed] Fixed an issue where the number value `-0.0` would lose its sign when
+  stored in Firestore.
+
+# 1.10.0
 - [feature] Implemented `Timestamp.valueOf()` so that `Timestamp` objects can be
   compared for relative ordering using the JavaScript arithmetic comparison
   operators (#2632).